Project/FilmService/requirements.txt
Django
mysqlclient
djangorestframework
requests
django-q